feat: update pagination handling across various API routes and schemas for consistency

+++ src/modules/class/schemas/class.schema.ts
@@ -26,8 +26,8 @@ export const classUpdateSchema = classFieldsSchema.partial().superRefine(validat
 export const classFilterSchema = z.object({
   search: z.string().optional(),
   status: z.enum(["DRAFT", "ACTIVE", "COMPLETED", "CANCELLED"]).optional(),
-  page: z.number().min(1).default(1),
-  pageSize: z.number().min(1).max(100).default(10),
+  page: z.number().int().min(1).default(1),
+  pageSize: z.number().int().min(1).max(100).default(10),
 });
 
 export type ClassCreate = z.infer<typeof classCreateSchema>;
